[not verified] VideoPress: Extract VideoThumbnail from VideoCard
