+++ projects/packages/videopress/src/client/admin/components/video-card/stories/VideoCard.mdx
@@ -12,13 +12,3 @@ React component that shows a card with details about the given video. It is poss
 It is composed of two simpler components: VideoThumbnail and VideoDetails components.
 
 <Story id="packages-videopress-video-card--default" />
-
--------
-
-# VideoThumbnailDropdown
-
-<Story id="packages-videopress-video-card--video-dropdown" />
-
-# VideoThumbnail
-
-<Story id="packages-videopress-video-card--video-thumbnail" />
