+++ projects/packages/videopress/src/client/admin/components/video-card/stories/index.tsx
@@ -5,11 +5,7 @@ import { action } from '@storybook/addon-actions';
 /**
  * Internal dependencies
  */
-import {
-	VideoCard as VideoCardComponent,
-	VideoThumbnail as VideoThumbnailComponent,
-	VideoThumbnailDropdown as VideoThumbnailDropdownComponent,
-} from '..';
+import { VideoCard as VideoCardComponent } from '..';
 import { postersArray, randomPoster } from '../../../mock';
 import Doc from './VideoCard.mdx';
 /**
@@ -51,32 +47,4 @@ _default.args = {
 	onUpdateThumbnailClick: action( 'onUpdateThumbnailClick' ),
 	onUpdateUpdatePrivacyClick: action( 'onUpdateUpdatePrivacyClick' ),
 	onDeleteClick: action( 'onDeleteClick' ),
-};
-
-const VideoThumbnailTemplate: ComponentStory< typeof VideoThumbnailComponent > = args => {
-	if ( args.thumbnail === 'none' ) {
-		args.thumbnail = null;
-	}
-
-	return <VideoThumbnailComponent { ...args } />;
-};
-
-export const VideoThumbnail = VideoThumbnailTemplate.bind( {} );
-VideoThumbnail.args = {
-	thumbnail: randomPoster(),
-	onUseDefaultThumbnail: action( 'onUseDefaultThumbnail' ),
-	onSelectFromVideo: action( 'onSelectFromVideo' ),
-	onUploadImage: action( 'onUploadImage' ),
-	editable: true,
-};
-
-const VideoThumbnailDropdownTemplate: ComponentStory<
-	typeof VideoThumbnailDropdownComponent
-> = VideoThumbnailDropdownComponent;
-
-export const VideoDropdown = VideoThumbnailDropdownTemplate.bind( {} );
-VideoDropdown.args = {
-	onUseDefaultThumbnail: action( 'onUseDefaultThumbnail' ),
-	onSelectFromVideo: action( 'onSelectFromVideo' ),
-	onUploadImage: action( 'onUploadImage' ),
 };
