+++ projects/packages/videopress/src/client/admin/components/video-thumbnail/stories/style.module.scss
@@ -1,0 +1,5 @@
+.decorator {
+	max-width: 500px;
+	min-height: 250px;
+	position: relative;
+}
